requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
alembic==1.14.0
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
asyncpg==0.30.0
attrs==24.2.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
click==8.1.7
deprecation==2.1.0
Flask==3.0.3
Flask-JWT-Extended==4.6.0
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
flask-supabase==0.2.1
frozenlist==1.5.0
gotrue==2.10.0
greenlet==3.1.1
greenletio==0.11.0
gunicorn==23.0.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.6
httpx==0.27.2
Hypercorn==0.17.3
hyperframe==6.0.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.6
MarkupSafe==3.0.2
multidict==6.1.0
packaging==24.1
postgrest==0.17.2
priority==2.0.0
propcache==0.2.0
psycopg2-binary==2.9.10
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
realtime==2.0.6
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
storage3==0.8.2
supabase==2.9.1
supafunc==0.6.2
typing_extensions==4.12.2
uvicorn==0.32.1
websockets==13.1
Werkzeug==3.1.1
wsproto==1.2.0
yarl==1.17.1
